requirements.txt
cx-Oracle==7.1.3
psycopg2==2.8.3
Pillow==6.0.0
opencv-python==4.1.0.25
pytesseract==0.2.7
pdf2jpg==0.0.9

